Add utils and reuse functions in torch and flax pipelines
